+++ CHANGES.md
@@ -11,6 +11,7 @@ Improvements 🙌:
 
 Bugfix 🐛:
  - Regression |  Share action menu do not work (#1647)
+ - verification issues on transition (#1555)
 
 Translations 🗣:
  -